+++ tests/test_field_parameters.py
@@ -1,3 +1,6 @@
+from datetime import date
+import sqlalchemy as sa
+from wtforms_alchemy import DateRange
 from wtforms.validators import NumberRange
 from tests import ModelFormTestCase
 
@@ -20,7 +23,16 @@ class TestFieldParameters(ModelFormTestCase):
         self.assert_default('test_column', "test")
 
     def test_min_and_max_info_attributes_generate_number_range_validator(self):
-        self.init(info={'min': 1, 'max': 100})
+        self.init(type_=sa.Integer, info={'min': 1, 'max': 100})
         validator = self.get_validator('test_column', NumberRange)
         assert validator.min == 1
         assert validator.max == 100
+
+    def test_min_and_max_info_attributes_generate_date_range_validator(self):
+        self.init(
+            type_=sa.Date,
+            info={'min': date(1990, 1, 1), 'max': date(2000, 1, 1)}
+        )
+        validator = self.get_validator('test_column', DateRange)
+        assert validator.min == date(1990, 1, 1)
+        assert validator.max == date(2000, 1, 1)