+++ tests/test_select_field_coerce.py
@@ -65,7 +65,7 @@ class TestSelectFieldCoerce(ModelFormTestCase):
 
     def test_unicode_coerces_values_to_unicode_strings(self):
         choices = [('1.0', '1.0'), ('2.0', '2.0')]
-        self.init(type_=sa.Unicode, info={'choices': choices})
+        self.init(type_=sa.Unicode(255), info={'choices': choices})
         form = self.form_class(MultiDict({'test_column': '2.0'}))
         assert form.test_column.data == u'2.0'
         assert isinstance(form.test_column.data, unicode)