+++ CHANGES.rst
@@ -3,12 +3,19 @@ Changelog
 
 Here you can see the full list of changes between each WTForms-Alchemy release.
 
-0.1.1 (2012-03-12)
+0.2 (2012-11-05)
 ^^^^^^^^^^^^^^^^^^
+
+- DateRange validator
+- SelectField with optgroup support
+
+
+0.1.1
+^^^^^
 
 - Added smart one-to-one and one-to-many relationship population
 
-0.1.0 (2012-03-11)
-^^^^^^^^^^^^^^^^^^
+0.1.0
+^^^^^
 
 - Initial public release